requirements.txt
aiogram==2.7
aiohttp==3.8.1
aioredis==2.0.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
et-xmlfile==1.1.0
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
multidict==5.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
